Revert "update to add the abstract class to invoke the init() in the constructor"
This reverts commit b1d9d1b40b835d3a2b653eec33ec2040be5ef847.
