+++ sharding-transaction/src/main/java/io/shardingsphere/transaction/manager/xa/XATransactionManagerSPILoader.java
@@ -50,6 +50,11 @@ public final class XATransactionManagerSPILoader {
         if (xaTransactionManagers.hasNext()) {
             log.warn("There are more than one transaction mangers existing, chosen first one by default.");
         }
+        try {
+            result.init();
+        } catch (Exception e) {
+            log.warn("Can not initialize " + result);
+        }
         return result;
     }
